requirements.txt
fastapi
uvicorn
PyYAML
requests
python-dateutil